+++ firebase-vertexai/src/main/kotlin/com/google/firebase/vertexai/type/Exceptions.kt
@@ -42,9 +42,7 @@ sealed class FirebaseVertexAIException(message: String, cause: Throwable? = null
             is com.google.ai.client.generativeai.common.ServerException ->
               ServerException(cause.message ?: "", cause.cause)
             is com.google.ai.client.generativeai.common.InvalidAPIKeyException ->
-              InvalidAPIKeyException(
-                cause.message ?: "",
-              )
+              InvalidAPIKeyException(cause.message ?: "")
             is com.google.ai.client.generativeai.common.PromptBlockedException ->
               PromptBlockedException(cause.response.toPublic(), cause.cause)
             is com.google.ai.client.generativeai.common.UnsupportedUserLocationException ->
@@ -91,7 +89,7 @@ class InvalidAPIKeyException(message: String, cause: Throwable? = null) :
 class PromptBlockedException(val response: GenerateContentResponse, cause: Throwable? = null) :
   FirebaseVertexAIException(
     "Prompt was blocked: ${response.promptFeedback?.blockReason?.name}",
-    cause
+    cause,
   )
 
 /**
@@ -121,7 +119,7 @@ class InvalidStateException(message: String, cause: Throwable? = null) :
 class ResponseStoppedException(val response: GenerateContentResponse, cause: Throwable? = null) :
   FirebaseVertexAIException(
     "Content generation stopped. Reason: ${response.candidates.first().finishReason?.name}",
-    cause
+    cause,
   )
 
 /**
@@ -141,10 +139,8 @@ class RequestTimeoutException(message: String, cause: Throwable? = null) :
 class InvalidLocationException(location: String, cause: Throwable? = null) :
   FirebaseVertexAIException("Invalid location \"${location}\"", cause)
 
-/**
- * The service is not enabled for this project. Visit the Firebase Console to enable it.
- */
-class ServiceDisabledException (message: String, cause: Throwable? = null) :
+/** The service is not enabled for this project. Visit the Firebase Console to enable it. */
+class ServiceDisabledException(message: String, cause: Throwable? = null) :
   FirebaseVertexAIException(message, cause)
 
 /** Catch all case for exceptions not explicitly expected. */